#2190 - Rename to RequireAem and deprecated AemCapabilityHelper
